Added table check for smoother deployment.

+++ db/migrate/20111010202847_create_issue_checklists.rb
@@ -1,14 +1,13 @@
 class CreateIssueChecklists < ActiveRecord::Migration
-  def self.up
+  def change
+    if ActiveRecord::Base.connection.table_exists? 'issue_checklists'
+      drop_table :issue_checklists
+    end
     create_table :issue_checklists do |t|
       t.boolean :is_done, :default => false
       t.string :subject
       t.integer :position, :default => 1
       t.references :issue, :null => false 
     end
-  end
-
-  def self.down
-    drop_table :issue_checklists
   end
 end
